requirements.txt
asgiref==3.11.1
async-timeout==5.0.1
certifi==2026.1.4
<<<<<<< HEAD
=======
channels==4.3.2
>>>>>>> development
charset-normalizer==3.4.4
Django==5.2.11
django-channels==0.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
oauthlib==3.3.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
